jcat: add --remove and --sort-fields arguments
With --remove we can remove fields, like lastUpdated, which can help
for example when comparing the metadata from different instances.

With --sort-fields we can choose which fields will be used to sort
the elements in the lists. For example, "name" will make the elements
sorted by what those elements contain in the "name" field. When the first
field is equal, the next one is used, and so on.

Finally, simplified the filters function, by using a helper function fix().