multirun.py: several improvements.
Change the names of certain arguments.

Has "hosts.txt" as the default file with the names of the hosts to run the command on.

Uses a more modern subprocess.run() instead of subprocess.check_output()

Shows the stderr when a command fails, with its error code and stdout that can be seen apart.

And some other cosmetic changes.