feat: add a folder for the delete spam scripts.
Added a README.md describing the scripts and how to use them,
Added a shell script to run the python script that gets the necessary DB connection parameters.
Cleaned up and fixed typo in the python script.
Added a new user to the default user list.
Added a week old filer to the deleted messages.